# docs/swagger.yaml
definitions:
  genprotos.GetAllUsers:
    properties:
      users:
        items:
          $ref: '#/definitions/genprotos.User'
        type: array
    type: object
  genprotos.User:
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
paths:
  /User/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete page
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete Successful
          schema:
            type: string
        "401":
          description: Error while Deleted
          schema:
            type: string
      summary: Delete User
      tags:
      - User
  /User/getall:
    get:
      consumes:
      - application/json
      description: GetAll page
      produces:
      - application/json
      responses:
        "200":
          description: GetAll Successful
          schema:
            $ref: '#/definitions/genprotos.GetAllUsers'
        "401":
          description: Error while GetAlld
          schema:
            type: string
      summary: GetAll User
      tags:
      - User
  /User/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.User'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById User
      tags:
      - User
  /User/login/{id}:
    get:
      consumes:
      - application/json
      description: GetById page
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetById Successful
          schema:
            $ref: '#/definitions/genprotos.User'
        "401":
          description: Error while GetByIdd
          schema:
            type: string
      summary: GetById User
      tags:
      - User
  /User/update/{id}:
    put:
      consumes:
      - application/json
      description: Update page
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: Update
        in: body
        name: Update
        required: true
        schema:
          $ref: '#/definitions/genprotos.User'
      produces:
      - application/json
      responses:
        "200":
          description: Update Successful
          schema:
            type: string
        "401":
          description: Error while created
          schema:
            type: string
      summary: Update User
      tags:
      - User
  /user/create:
    post:
      consumes:
      - application/json
      description: Create page
      parameters:
      - description: Create
        in: body
        name: Create
        required: true
        schema:
          $ref: '#/definitions/genprotos.User'
      produces:
      - application/json
      responses:
        "200":
          description: Create Successful
          schema:
            type: string
        "401":
          description: Error while Created
          schema:
            type: string
      summary: Create User
      tags:
      - User
swagger: "2.0"